Add unique password generation and tests
Modified the PasswordGenerator class to generate unique passwords and also store them. A new PHPUnitUtil class was created to facilitate testing of the new functionality. Further, migration files were added to handle the storage of passwords in the database. New tests assure non-repetitive passwords and check runtime limits for uniqueness.